add password recovery entities
