+++ entity/controllers/password_recovery.go
@@ -1,0 +1,6 @@
+package controllers
+
+type PasswordRecovery struct {
+	Token    string `json:"token" binding:"required,validateRecoveryToken"`
+	Password string `json:"password" binding:"required,validatePassword"`
+}
